+++ src/CleanRegex/Internal/Match/Stream/Base/MatchStream.php
@@ -4,7 +4,6 @@ namespace TRegx\CleanRegex\Internal\Match\Stream\Base;
 use TRegx\CleanRegex\Internal\Match\Stream\ListStream;
 use TRegx\CleanRegex\Internal\Match\Stream\SubjectStreamRejectedException;
 use TRegx\CleanRegex\Internal\Match\Stream\Upstream;
-use TRegx\CleanRegex\Internal\Match\UserData;
 use TRegx\CleanRegex\Internal\Message\SubjectNotMatched\FirstMatchMessage;
 use TRegx\CleanRegex\Internal\Model\DetailObjectFactory;
 use TRegx\CleanRegex\Internal\Model\FalseNegative;
@@ -22,20 +21,17 @@ class MatchStream implements Upstream
     private $stream;
     /** @var Subject */
     private $subject;
-    /** @var UserData */
-    private $userData;
     /** @var MatchAllFactory */
     private $allFactory;
     /** @var DetailObjectFactory */
     private $detailObjects;
 
-    public function __construct(StreamBase $stream, Subject $subject, UserData $userData, MatchAllFactory $allFactory)
+    public function __construct(StreamBase $stream, Subject $subject, MatchAllFactory $allFactory)
     {
         $this->stream = $stream;
         $this->subject = $subject;
-        $this->userData = $userData;
         $this->allFactory = $allFactory;
-        $this->detailObjects = new DetailObjectFactory($subject, $userData);
+        $this->detailObjects = new DetailObjectFactory($subject);
     }
 
     protected function entries(): array
@@ -49,8 +45,7 @@ class MatchStream implements Upstream
             $this->tryFirstKey(),
             1,
             new GroupPolyfillDecorator(new FalseNegative($this->stream->first()), $this->allFactory, 0),
-            $this->allFactory,
-            $this->userData);
+            $this->allFactory);
     }
 
     private function tryFirstKey(): int