+++ test/Fakes/CleanRegex/Match/Details/ThrowDetail.php
@@ -121,16 +121,6 @@ class ThrowDetail implements Detail
         throw $this->fail();
     }
 
-    public function setUserData($userData): void
-    {
-        throw $this->fail();
-    }
-
-    public function getUserData()
-    {
-        throw $this->fail();
-    }
-
     public function __toString(): string
     {
         throw $this->fail();